requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
python-dotenv==1.0.0
PyMySQL==1.1.0
line-bot-sdk==3.5.0
google-generativeai==0.3.2
azure-cognitiveservices-speech==1.34.0
Pillow==10.0.1
Werkzeug==2.3.7
gunicorn==21.2.0